Move custom entities to their own package
