+++ v1_10_R1/src/main/java/net/aufdemrand/denizen/nms/impl/entities/CraftItemProjectile_v1_10_R1.java
@@ -1,4 +1,4 @@
-package net.aufdemrand.denizen.nms.impl;
+package net.aufdemrand.denizen.nms.impl.entities;
 
 import net.aufdemrand.denizen.nms.interfaces.ItemProjectile;
 import org.bukkit.craftbukkit.v1_10_R1.CraftServer;
